requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.2.4
django-bootstrap4==3.0.1
django-chartjs==2.2.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
